6.21.5 Fix code review comments - OffersCardsListWithActive в отдельную папку и добавил тесты.
